+++ src/main/java/github/anandb/netbeans/mcp/EditorToolProvider.java
@@ -102,7 +102,7 @@ public class EditorToolProvider {
         required.add("line");
 
         mcpTools.registerTool(
-                "open_file_at_line",
+                "open_pos",
                 "Opens a file at the specified line number in the editor. Only works for files within the current project. " +
                 "Call when you see phrases like 'Show me where this happens in the code, or 'open the files where we use this'" +
                 "'Where all do we have this kind of pattern', 'Show me the code where XXX'",